move calculateAllGroupGuideRates into BlackoilWellModelGuideRates
